Abandoning branch; Godot cannot do anything we need it to.
The idea of exporting a PNG with embedded metadata was appealing, but
Godot's PNG support is very limited and does not extend to writing/reading
PNG metadata.

This brought about the question of, 'if we can't import/export creatures
using PNGs, is there any value?' and there MIGHT still be value if we had
something like, letting the player pick a few backgrounds or things like
that.

As far as letting the player pick a few backgrounds -- the most intuitive
way to implement this would be to render it to a viewport with a
background color. Unfortunately Godot does not support background colors for 2D
viewports (https://github.com/godotengine/godot-proposals/issues/252).

A second way to implement this would be to create a blank image with a
color behind it, and draw the creature's image on top. This is not
supported by the Image API. It does provide 'get_pixel' and 'set_pixel'
methods so I guess we could do it pixel-by-pixel at the software level,
but the idea of writing a software-level shader code seems slow and
frankly insane.

A third way to implement this would be to, I guess, render something on
the screen and mess with the screen capture I guess? I don't really care.
This isn't worth it, it provides almost no benefit to anybody. People can
just use screen captures.